% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moose12687}
\alias{moose12687}
\title{Radio telemetry data for one Minnesota moose (ID #12687) from 2013 and 2014.}
\format{A data frame with 15158 rows and 13 variables:
\describe{
  \item{CollarID}{The unique identifier for the GPS collar.}
  \item{datetime}{The timestamp for each record.}
  \item{stratum}{The stratum identifier. One stratum is included for each
  observed location (presence==1) and its associated available locations
  generated by the random movement paths (presence==0).}
  \item{X}{The X coordinate (meters; NAD1983 UTM Zone 15N).}
  \item{Y}{The Y coordinate (meters; NAD1983 UTM Zone 15N).}
  \item{step}{Step length (mm) from the previous location. Original step
  length in meters was divided by 1000 to make coefficients similar within
  the step-selection function. See Fieberg et al. (In Review) for more
  information.}
  \item{bearing}{The bearing relative to true north.}
  \item{presence}{Binary indicator for each observed location (presence==0)
  and its associated available locations generated by the random movement
  paths (presence==0).}
  \item{decid50}{Percent cover of deciduous forest from 2011 NLCD (Jin et
  al. 2013).}
  \item{mixed50}{Percent cover of mixedwood forest from 2011 NLCD (Jin et
  al. 2013).}
  \item{conif50}{Percent cover of conifer forest from 2011 NLCD (Jin et
  al. 2013).}
  \item{treedwet50}{Percent cover of treed wetlands from 2011 NLCD (Jin
  et al. 2013).}
  \item{year}{The year for each point (2013 or 2014).}
}}
\source{
Original MN moose data was derived from Street et al. (2016)
  (\url{http://doi.org/10.1007/s10980-016-0372-z}). National Land Cover Data
  was derived from: Jin et al. (2013) A comprehensive change detection
  method for updating National Land Cover Dataset to circa 2011, Remote
  Sensing of the Environment. A full archive of the data and
  code necessary to replicate the manuscript can be found at
  \url{http://doi.org/10.13020/D6T590}
}
\usage{
moose12687
}
\description{
A dataset including presence locations and randomly-generated available
locations of a moose in Minnesota in 2013 and 2014. Data was derived
from Street et al. (2016).
}
\seealso{
uhcsimstrat
}
\keyword{datasets}

