% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhcdatasimulator.R
\name{uhcdatasimulator}
\alias{uhcdatasimulator}
\title{uhcdatasimulator}
\usage{
uhcdatasimulator(nused, navail, betas, corx, ntemp, scenario, example)
}
\arguments{
\item{nused}{The number of used locations in training/test data set}

\item{navail}{The number of background locations in training/test data set}

\item{betas}{The vector of length 2 for *true* probability of use}

\item{corx}{The correlation between elevation and precipitation in
training/test dataset. For missing predictor example only.}

\item{ntemp}{A large number of available points.}

\item{scenario}{The specific scenario as described in the manuscript.
Options include"corr00", "corrNN", "corrPN", and "T".}

\item{example}{The name of the example. Options include "missing predictor"
or "non-linear".}
}
\value{
A dataframe of simulated data
}
\description{
\code{uhcdatasimulator} simulates the data used in Fieberg et al. (In Review)
}
\details{
This is a function that creates a dataframe based on the example and scenario
chosen from the manuscript (Fieberg et al. In Review). In the first example,
("missing predictor") the distribution of a species was related to elevation
and precipitation. In the second example ("non-linear"), the distribution of
a species is non-linearly related to temperature.
}
\examples{
# Simulate data for the non-linear example
new.data <- uhcdatasimulator(nused = 100,
   navail = 10000,
   betas = c(2,-1),
   ntemp = 1000000,
   scenario = "T",
   example = "non-linear")
}

