% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhcdenscalc.R
\name{uhcdenscalc}
\alias{uhcdenscalc}
\title{uhcdenscalc}
\usage{
uhcdenscalc(rand_sims, dat, avail, gridsize = 500)
}
\arguments{
\item{rand_sims}{The characteristics associated with the simulated locations
in the test data set}

\item{dat}{The characteristics associated with observed locations in the
test data set}

\item{avail}{The characteristics associated with available locations in the
test data set}

\item{gridsize}{The size of grid for estimating the kernel density estimate
(kde). Default value is 500.}
}
\value{
A list of density estimates associated with the simulated locations,
(densrand), the observed locations (densdat), and the available locations
(densavail)
}
\description{
\code{uhcdenscalc} calculates kernal density estimates to create UHC plots.
}
\details{
\code{uhcdenscalc} calculates density estimates for the environmental
characteristics associated with the observed and available locations in the
test data and also those associated with the randomly chosen locations
generated by the \code{uhcsim} or \code{uhcsimstrat} functions.
}
\examples{
# Simulate training data for the non-linear example
nonlinear.train <- uhcdatasimulator(nused = 100,
    navail = 10000,
    betas = c(2,-1),
    ntemp = 1000000,
    example = "non-linear")

# Simulate test data for the non-linear example
nonlinear.test <- uhcdatasimulator(nused = 100,
    navail = 10000,
    betas = c(2,-1),
    ntemp = 1000000,
    example = "non-linear")

# Fit GLM with quadratic relationship
train.correct <- glm(y~temp + I(temp^2),
   family = binomial,
   data = nonlinear.train)

# Fit GLM with linear (misspecified) relationship
train.misspec <- glm(y~temp,
   family = binomial,
   data = nonlinear.train)

# Simulate data for quadratic model
xhat.correct <- uhcsim(nsims = 1000,
   nused_test = 100,
   xmat = model.matrix(y~temp + I(temp^2), data = nonlinear.test)[,-1],
   fit_rsf = train.correct,
   z = as.matrix(nonlinear.test[,"temp"]))

# Simulate data for linear (misspecified) model
xhat.misspec <- uhcsim(nsims = 1000,
   nused_test = 100,
   xmat = as.matrix(model.matrix(y~temp, data = nonlinear.test)[,2]),
   fit_rsf = train.misspec,
   z = as.matrix(nonlinear.test[,"temp"]))

# Get density estimates for quadratic model
denshats.correct <- uhcdenscalc(rand_sims = xhat.correct[,,1],
   dat = subset(nonlinear.test, y==1, select="temp"),
   avail = subset(nonlinear.test, y==0, select="temp"))

# Get density estimates for linear (misspecified) model
denshats.misspec <- uhcdenscalc(rand_sims = xhat.misspec[,,1],
   dat = subset(nonlinear.test, y==1, select="temp"),
   avail = subset(nonlinear.test, y==0, select="temp"))
}
\seealso{
Full archive of the data and code necessary to replicate the
manuscript at \url{http://doi.org/10.13020/D6T590}.
}
