% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhcdiffdensplot.R
\name{uhcdiffdensplot}
\alias{uhcdiffdensplot}
\title{uhcdiffdensplot}
\usage{
uhcdiffdensplot(densdat, densrand, xl = NULL)
}
\arguments{
\item{densdat}{The kernel density estimates of observed points in the test
data set}

\item{densrand}{The predicted kernel density estimates of simulated data
points}

\item{xl}{The x-axis limits (can be user supplied)}
}
\value{
A plot with the mean (black line) and upper and lower bounds of a
95% simulation envelope (red dashed lines) of the difference
between the predicted density estimates and the density estimates for the
presence locations.
}
\description{
\code{uhcdiffdensplot} plots an alternative UHC plot with a simulation
envelope for \eqn{f^U(z) - \hat{f}^u(z)}
}
\examples{
# Simulate training data for the non-linear example
nonlinear.train <- uhcdatasimulator(nused = 100,
    navail = 10000,
    betas = c(2,-1),
    ntemp = 1000000,
    example = "non-linear")

# Simulate test data for the non-linear example
nonlinear.test <- uhcdatasimulator(nused = 100,
    navail = 10000,
    betas = c(2,-1),
    ntemp = 1000000,
    example = "non-linear")

# Fit GLM with quadratic relationship
train.correct <- glm(y~temp + I(temp^2),
   family = binomial,
   data = nonlinear.train)

# Fit GLM with linear (misspecified) relationship
train.misspec <- glm(y~temp,
   family = binomial,
   data = nonlinear.train)

# Simulate data for quadratic model
xhat.correct <- uhcsim(nsims = 1000,
   nused_test = 100,
   xmat = model.matrix(y~temp + I(temp^2), data = nonlinear.test)[,-1],
   fit_rsf = train.correct,
   z = as.matrix(nonlinear.test[,"temp"]))

# Simulate data for linear (misspecified) model
xhat.misspec <- uhcsim(nsims = 1000,
   nused_test = 100,
   xmat = as.matrix(model.matrix(y~temp, data = nonlinear.test)[,2]),
   fit_rsf = train.misspec,
   z = as.matrix(nonlinear.test[,"temp"]))

# Get density estimates for quadratic model
denshats.correct <- uhcdenscalc(rand_sims = xhat.correct[,,1],
   dat = subset(nonlinear.test, y==1, select="temp"),
   avail = subset(nonlinear.test, y==0, select="temp"))

# Get density estimates for linear (misspecified) model
denshats.misspec <- uhcdenscalc(rand_sims = xhat.misspec[,,1],
   dat = subset(nonlinear.test, y==1, select="temp"),
   avail = subset(nonlinear.test, y==0, select="temp"))

# Create an alternate UHC plot for the quadratic model
uhcdiffdensplot(densdat = denshats.correct$densdat,
   densrand = denshats.correct$densrand)

# Create an alternate UHC plot for the linear (misspecified) model
uhcdiffdensplot(densdat = denshats.misspec$densdat,
   densrand = denshats.misspec$densrand)
}
\seealso{
Full archive of the data and code necessary to replicate the
manuscript at \url{http://doi.org/10.13020/D6T590}.
}

