% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhcplots.R
\docType{package}
\name{uhcplots}
\alias{uhcplots}
\alias{uhcplots-package}
\title{Used-Habitat Calibration Plots (UHC Plots)}
\description{
Used-Habitat Calibration Plots (UHC Plots)
}
\section{Details}{


Used-Habitat Calibration plots are a new method for model calibration that
can be applied across a spectrum of species distribution models (SDMs)
(Fieberg et al. In revision for Ecology Letters).
}

\section{See also}{



Fieberg, Forester, Street, Johnson, ArchMiller, and Matthiopoulos.
In Revision. Species distribution models: predictive snipers or shots in the
dark? Ecology Letters.

Full code to reproduce the paper at: <http://doi.org/10.13020/D6T590>
}

