% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhcsim.R
\name{uhcsim}
\alias{uhcsim}
\title{uhcsim}
\usage{
uhcsim(nsims, nused_test, xmat, fit_rsf, z)
}
\arguments{
\item{nsims}{The number of simulations (M) used to create the UHC plot.}

\item{nused_test}{The number of used locations in the test data set.}

\item{xmat}{A matrix of predictor variables in the test data.}

\item{fit_rsf}{The fitted logistic regression model object}

\item{z}{A vector or matrix of (used & available) environmental
characteristics in the test data set.}
}
\value{
An array of dimensions nsims x nused_test x p.
}
\description{
\code{uhcsim} samples, randomly, locations from non-stratified
test data.
}
\details{
This function samples, randomly, locations from non-stratified test data and
returns an array of dimension nsims x nused_test x p (where p
is the number of predictors to be validated)
}
\examples{
# Simulate training data for the non-linear example
nonlinear.train <- uhcdatasimulator(nused = 100,
   navail = 10000,
   betas = c(2,-1),
   ntemp = 1000000,
   example = "non-linear")

# Simulate test data for the non-linear example
nonlinear.test <- uhcdatasimulator(nused = 100,
   navail = 10000,
   betas = c(2,-1),
   ntemp = 1000000,
   example = "non-linear")

# Fit GLM with quadratic relationship
train.correct <- glm(y~temp + I(temp^2),
   family = binomial,
   data = nonlinear.train)

# Fit GLM with linear (misspecified) relationship
train.misspec <- glm(y~temp,
   family = binomial,
   data = nonlinear.train)

# Simulate data for quadratic model
xhat.correct <- uhcsim(nsims = 1000,
   nused_test = 100,
   xmat = model.matrix(y~temp + I(temp^2), data = nonlinear.test)[,-1],
   fit_rsf = train.correct,
   z = as.matrix(nonlinear.test[,"temp"]))

# Simulate data for linear (misspecified) model
xhat.misspec <- uhcsim(nsims = 1000,
   nused_test = 100,
   xmat = as.matrix(model.matrix(y~temp, data = nonlinear.test)[,2]),
   fit_rsf = train.misspec,
   z = as.matrix(nonlinear.test[,"temp"]))
}
\seealso{
Full archive of the data and code necessary to replicate the
manuscript at \url{http://doi.org/10.13020/D6T590}.
}

