% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhcsimXvalid.R
\name{uhcsimXvalid}
\alias{uhcsimXvalid}
\title{uhcsimXvalid}
\usage{
uhcsimXvalid(data_sim, k_folds, nsims, nused, navail, ncovariates, model_form,
  z_colnames, xmat_colnames)
}
\arguments{
\item{data_sim}{The data set to be used in cross-validation}

\item{k_folds}{The number of bins to split data into}

\item{nsims}{The number of simulations (M) used to create the UHC plot.}

\item{nused}{The number of used locations in the data set.}

\item{navail}{The number of available locations in the data set (ie not used)}

\item{ncovariates}{The number of covariates or p in the model}

\item{model_form}{The model in the form: "y~x1+x2+...+xp" (including
quotations)}

\item{z_colnames}{The column names of the environmental characteristics to be
plotted with \code{uhcdensplot}. Takes the form: c("x1","x2",...,"xp")}

\item{xmat_colnames}{The column names of the environmental characteristics
that were used as predictors in the model_form. Takes the form:
c("x1","x2",...,"xp")}
}
\value{
An array of dimensions nsims x nused_test x p.
}
\description{
\code{uhcsimXvalid} samples, randomly, locations from non-stratified
test data using k-fold cross validation.
}
\details{
This function samples, randomly, locations from non-stratified test data and
returns an array of dimension nsims x nused_test x p (where p
is the number of predictors to be validated)
}
\examples{

}
\seealso{
Full archive of the data and code necessary to replicate the
manuscript at \url{http://doi.org/10.13020/D6T590}.
}

