% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhcdatasimulator.R
\name{uhcdatasimulator}
\alias{uhcdatasimulator}
\title{uhcdatasimulator}
\usage{
uhcdatasimulator(nused, navail, betas, corx, ntemp, example)
}
\arguments{
\item{nused}{The number of used locations in training/test data set}

\item{navail}{The number of background locations in training/test data set}

\item{betas}{The vector of length 2 for *true* probability of use}

\item{corx}{The correlation between elevation and precipitation in
training/test dataset. For missing predictor example only.}

\item{ntemp}{A large number of available points.}

\item{example}{The name of the example. Options include "missing predictor"
or "non-linear".}
}
\value{
A dataframe of simulated data
}
\description{
\code{uhcdatasimulator} simulates the data used in Fieberg et al. (In Review)
}
\details{
This is a function that creates a dataframe based on the example
chosen from the manuscript (Fieberg et al. In Review). In the first example,
("missing predictor") the distribution of a species is related to elevation
(\eqn{x_1}) and precipitation (\eqn{x_2}), where \eqn{x_1} and \eqn{x_2}
are normally distributed with mean 0 and variance 4. We considered 3
different data-generating scenarios in which we varied the correlation of
\eqn{x_1,x_2} (\code{corx}) in the training and test data sets.
In the second example ("non-linear"), the distribution of
a species is non-linearly related to temperature (\eqn{x_3 \sim N(0,4)}).
}
\examples{
# Simulate training or test data for the non-linear example
nonlinear.data <- uhcdatasimulator(nused = 100,
   navail = 10000,
   betas = c(2,-1),
   ntemp = 1000000,
   example = "non-linear")

# Simulate training or test data for the missing predictor example
   #   Where corr(x1,x2) = 0
missingpredictor.0.data <- uhcdatasimulator(nused = 100,
   navail = 10000,
   betas = c(0.5,-1),
   corx = 0,
   ntemp = 1000000,
   example = "missing predictor")

   #   Where corr(x1,x2) = -0.3
missingpredictor.N.data <- uhcdatasimulator(nused = 100,
   navail = 10000,
   betas = c(0.5,-1),
   corx = -0.3,
   ntemp = 1000000,
   example = "missing predictor")

   #   Where corr(x1,x2) = 0.3
missingpredictor.P.data <- uhcdatasimulator(nused = 100,
   navail = 10000,
   betas = c(0.5,-1),
   corx = 0.3,
   ntemp = 1000000,
   example = "missing predictor")
}
\seealso{
Full archive of the data and code necessary to replicate the
manuscript at \url{http://doi.org/10.13020/D6T590}.
}

