% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhcplots.R
\name{uhcplots}
\alias{uhcplots}
\title{Package development tools for R.
# Used-Habitat Calibration Plots (UHC Plots)}
\description{
Package development tools for R.
# Used-Habitat Calibration Plots (UHC Plots)
}
\section{See also}{


Fieberg, et al. (In Revision)
}

